% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccfun.R
\docType{package}
\name{ccfun}
\alias{ccfun}
\alias{ccfun-package}
\title{ccfun: A package for manipulating the HIC critical care data}
\description{
The foo package provides three categories of important functions:
foo, bar and baz.
}
\section{Foo functions}{

The foo functions ...
Created by following http://r-pkgs.had.co.nz/man.html
You'll note that there is no function here (see \code{NULL}) below.
Moreover you have to explicitly define \code{@doctype} and \code{@name}
}

