% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen.R
\name{choose_first_nonmissing}
\alias{choose_first_nonmissing}
\title{Choose first non-missing}
\usage{
choose_first_nonmissing(colvalues)
}
\arguments{
\item{colvalues}{columns to combine}
}
\description{
Choose first non-missing
}
