% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_ICNARC.R
\name{chop.icnarc.code}
\alias{chop.icnarc.code}
\title{Parses ICNARC diagnostic codes}
\usage{
chop.icnarc.code(dt, dc.col, prefix = NULL)
}
\arguments{
\item{dt}{data.table containing diagnostic codes to be parsed}

\item{dc.col}{column name within data table with diagnostic codes}

\item{prefix}{new diagnostic component column prefix after parsing
Split ICNARC text code, and convert to numeric}
}
\description{
Split text codes by level and then label as per ICNARC codes
}

