% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_gcs.R
\name{clean_gcs}
\alias{clean_gcs}
\title{Cleans GCS by replacing with NA where concurrently sedated}
\usage{
clean_gcs(data, gcs_ = "gcs", sedatives_ = NULL, verbose = FALSE)
}
\arguments{
\item{gcs_}{GCS}

\item{sedatives_}{List of field names for sedatives; if null then uses the ITEM_REF dictionary}

\item{dt}{data.table containing physiology data}
}
\description{
Cleans GCS score
}
