% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NHIC_utilities.R
\name{convert_numeric}
\alias{convert_numeric}
\title{Convert those columns to numeric}
\usage{
convert_numeric(dt, dict_ = dict)
}
\arguments{
\item{dt}{data.table to convert}

\item{dict_}{NHIC dictionary from YAML file}
}
\description{
Convert those columns to numeric
}
