% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relabel_cols.R
\name{dict_translate}
\alias{dict_translate}
\title{Relabels a wide data2d data.table}
\usage{
dict_translate(dict, value.a, label.in, label.out)
}
\arguments{
\item{dict}{defaults to ITEM_REF}

\item{label.in}{name for the item to be translated}

\item{label.out}{name of the item to be used for the new column name}

\item{dt}{data.table to be relabelled}
}
\value{
0 (if the function completes)
}
\description{
Relabels the columns of a wide data2d data.table (as generated by
the selectTable function in ccdata). Because this uses the
data.table::setnames function then the name changes happen 'in place'
and nothing is returned
}
\examples{
# use the ITEM_REF dictionary (not run)
# relabel_cols(tdt, "dataItem", "shortName")
# use a custom dictionary (not run)
# relabel_cols(tdt, "dataItem", "shortName", dict=analysis.dict)
}

