% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NHIC_utilities.R
\name{find_NHIC}
\alias{find_NHIC}
\title{Look for possible matches in NHIC dict}
\usage{
find_NHIC(k, v, dict_ = dict)
}
\arguments{
\item{k}{key}

\item{v}{value}

\item{dict_}{NHIC dictionary from YAML file}
}
\description{
Look for possible matches in NHIC dict
}

