% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_apache_aki.R
\name{gen_apache_aki}
\alias{gen_apache_aki}
\title{Generates the APACHE Acute Kidney Injury score}
\usage{
gen_apache_aki(dt, input, with_ckd = F, output = NULL)
}
\arguments{
\item{dt}{data.table containing physiology data}

\item{input}{a vector of numeric data}

\item{with_ckd}{a logical vector, preexisting chronic renal function impairement

 @export}

\item{output}{Column Name for the result of computation}
}
\description{
Generates the APACHE aki score;
}

