% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_apache_gcs.R
\name{gen_apache_gcs}
\alias{gen_apache_gcs}
\title{Generates the APACHE Glasgow Coma Scale score}
\usage{
gen_apache_gcs(dt, input, output = NULL)
}
\arguments{
\item{dt}{data.table containing physiology data}

\item{input}{a vector of numeric data

 @export}

\item{output}{Column Name for the result of computation}
}
\description{
Generates the APACHE gcs score;
}

