% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_apache_main.R
\name{gen_apache_main}
\alias{gen_apache_main}
\title{Generates the APACHE score from the APACHE sub-categories}
\usage{
gen_apache_main(mylist, output = NULL)
}
\arguments{
\item{output}{Column Name for the result of computation}

\item{dt}{data.table containing physiology data}

\item{input}{a list of numeric data

 @export}
}
\description{
Generates the APACHE score using the preexisting APACHE sub-categories
}

