% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_apache_map.R
\name{gen_apache_map}
\alias{gen_apache_map}
\title{Generates the APACHE Arterial Pressure score}
\usage{
gen_apache_map(dt, input, output = NULL)
}
\arguments{
\item{dt}{data.table containing physiology data}

\item{input}{Arterial Pressure}

\item{output}{Column name for the result of computation}
}
\description{
Generates the APACHE Arterial Pressure score;
}
\examples{
# system.time(gen_apache_map(ddata, temp_ = map))
# table(ddata$apache_map, useNA="always")
# ddata[map > 40][sample(nrow(ddata[map > 40]),20), .(map)]
 @export
}

