% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_apache_mf.R
\name{gen_apache_mf}
\alias{gen_apache_mf}
\title{Generates the APACHE Metabolic Failure score}
\usage{
gen_apache_mf(dt, hco3, ph, output = NULL)
}
\arguments{
\item{dt}{data.table containing physiology data}

\item{hco3}{a vector of numeric data}

\item{ph}{a vector of numeric data

 @export}

\item{output}{Column name for the result of computation}
}
\description{
Generates the APACHE Metabolic Failure score;
}

