% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_apache_Na.R
\name{gen_apache_Na}
\alias{gen_apache_Na}
\title{Generates the APACHE Sodium score}
\usage{
gen_apache_Na(dt, input, output = NULL)
}
\arguments{
\item{dt}{data.table containing physiology data}

\item{input}{a vector of numeric data

 @export}

\item{output}{Column name for the result of computation}
}
\description{
Generates the APACHE Sodium score;
}

