% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_apache_rf.R
\name{gen_apache_rf}
\alias{gen_apache_rf}
\title{Generates the APACHE Respiratory Failure score}
\usage{
gen_apache_rf(dt, input, output = NULL, pao2, paco2)
}
\arguments{
\item{dt}{data.table containing physiology data}

\item{input}{Respiratory Failure Component}

\item{output}{Column name for the result of computation}

\item{pao2}{a vector of numeric data}

\item{paco2}{a vector of numeric data

 @export}
}
\description{
Generates the APACHE Respiratory Failure score;
}

