% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_apache_rr.R
\name{gen_apache_rr}
\alias{gen_apache_rr}
\title{Generates the APACHE Respiratory Rate score}
\usage{
gen_apache_rr(dt, input, output = NULL)
}
\arguments{
\item{dt}{data.table containing physiology data}

\item{input}{Respiratory Rate (RR), integer or numeric}

\item{output}{Column name for the result of computation}
}
\description{
Generates the APACHE Respiratory Rate score; requires at least one respiratory rate input
}
\examples{
# system.time(gen_apache_rr(ddata, input = RR))
# table(ddata$apache_rr, useNA="always")
# ddata[RR < 45][sample(nrow(ddata[RR < 45]),20), .(RR)]
 @export
}

