% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_apache_temp.R
\name{gen_apache_temp}
\alias{gen_apache_temp}
\title{Generates the APACHE Temperature score}
\usage{
gen_apache_temp(dt, input, output = NULL)
}
\arguments{
\item{dt}{data.table containing physiology data}

\item{input}{Temperature}

\item{output}{Column name for the result of computation}
}
\description{
Generates the APACHE Temperature score;
}
\examples{
# system.time(gen_apache_temp(ddata, temp_ = Temperature))
# table(ddata$apache_temp, useNA="always")
# ddata[Temperature > 40][sample(nrow(ddata[Temperature > 40]),20), .(Temperature)]
 @export
}

