% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_apache_wbc.R
\name{gen_apache_wbc}
\alias{gen_apache_wbc}
\title{Generates the APACHE White Blood Cell Count score}
\usage{
gen_apache_wbc(dt, input, output = NULL)
}
\arguments{
\item{dt}{data.table containing physiology data}

\item{input}{a vector of numeric data

 @export}

\item{output}{Column Name for the result of computation}
}
\description{
Generates the APACHE wbc score;
}

