% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_neuro.R
\name{gen_neuro}
\alias{gen_neuro}
\title{Generates the neuro score for various scoring systems}
\usage{
gen_neuro(dt, score, gcs_, hours_, id_)
}
\arguments{
\item{dt}{data.table containing physiology data}

\item{score}{score that the nero component is to be calculated for}

\item{gcs_}{GCS total score}

\item{hours_}{hours since ICU admission}

\item{id_}{unique id for each patient}
}
\description{
Generates the neuro score; requires GCS
}
\examples{
# gen_neuro(ddata,sofa, gcs_ = gcs)
# table(ddata$sofa_n, useNA="always")
# ddata[gcs<=15][sample(nrow(ddata[gcs<=15]),20), .(gcs, sofa_n)]
}
