% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_sofa_c.R
\name{gen_sofa_c}
\alias{gen_sofa_c}
\title{Generates the SOFA cardiovascular score}
\usage{
gen_sofa_c(dt, map_, norad_, adr_ = NULL, dopa_ = NULL, dobu_ = NULL,
  vasop_ = NULL)
}
\arguments{
\item{dt}{data.table containing physiology data}

\item{map_}{Mean Arterial Pressure}

\item{norad_}{noradrenaline dose in mcg/kg/min}
}
\description{
Generates the SOFA CVS score; requires MAP and norad and optionally
handles other vasopressors
}
\examples{
# system.time(gen_sofa_c(ddata, map_=map, norad_=rx_norad, adr_=rx_adre, vasop_=rx_vasopr, dopa_=rx_dopa, dobu_=rx_dobu) )
# table(ddata$sofa_c, useNA="always")
# ddata[rx_adre<0.1][sample(nrow(ddata[rx_adre<0.1]),20), .(map, rx_norad, rx_adre, rx_dopa, rx_vasopr, rx_dobu, sofa_c)]
}
