% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_sofa_h.R
\name{gen_sofa_h}
\alias{gen_sofa_h}
\title{Generates the SOFA haematology score}
\usage{
gen_sofa_h(dt, platelets_)
}
\arguments{
\item{dt}{data.table containing physiology data}

\item{platelets_}{platelets}
}
\description{
Generates the SOFA haematology score; requires platelets
}
\examples{
# gen_sofa_h(ddata, platelets_ = platelets)
# ddata[platelets<400][sample(nrow(ddata[platelets<400]),20), .(platelets, sofa_h)]
# table(ddata$sofa_h, useNA="always")
}

