% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_sofa_k.R
\name{gen_sofa_k}
\alias{gen_sofa_k}
\title{Generates the SOFA renal score}
\usage{
gen_sofa_k(dt, creat_, uvol24h_ = NULL)
}
\arguments{
\item{dt}{data.table containing physiology data}

\item{creat_}{creatinine (micromol/l)}

\item{uvol24h_}{24 hourly urine volume}

\item{rrt_}{current use of renal replacement therapy}
}
\description{
Generates the SOFA renal score; requires creatinine and optionally urine
}
\examples{
# gen_sofa_k(ddata, creat_ = creatinine)
# table(ddata$sofa_k, useNA="always")
# ddata[creatinine < 500][sample(nrow(ddata[creatinine < 500]),20), .(creatinine, sofa_k)]
}
