% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_sofa_l.R
\name{gen_sofa_l}
\alias{gen_sofa_l}
\title{Generates the SOFA liver score}
\usage{
gen_sofa_l(dt, bili_)
}
\arguments{
\item{dt}{data.table containing physiology data}

\item{bili_}{bilirubin}
}
\description{
Generates the SOFA liver score; requires bilirubin
}
\examples{
# gen_sofa_l(ddata, bilirubin_ = bilirubin)
# table(ddata$sofa_l, useNA="always")
# ddata[bilirubin>100][sample(nrow(ddata[bilirubin>100]),20), .(bilirubin, sofa_l)]
}
