% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_sofa_n.R
\name{gen_sofa_n}
\alias{gen_sofa_n}
\title{Generates the SOFA neuro score}
\usage{
gen_sofa_n(dt, gcs_)
}
\arguments{
\item{dt}{data.table containing physiology data}

\item{gcs_}{GCS}
}
\description{
Generates the SOFA neuro score; requires GCS
}
\examples{
# gen_sofa_n(ddata, gcs_ = gcs)
# table(ddata$sofa_n, useNA="always")
# ddata[gcs<=15][sample(nrow(ddata[gcs<=15]),20), .(gcs, sofa_n)]
}
