% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspect_item.R
\name{inspect_factor}
\alias{inspect_factor}
\title{Inspects NHIC data items}
\usage{
inspect_factor(d)
}
\arguments{
\item{d}{data generated from the ccdata::readItems() function}

\item{f.dict}{field dictionary generated from ITEM_REF.yml
via ccfun::load_ccDict.R}

\item{...}{args to helper functions (e.g. palette="Greens" to ggMMplot)}
}
\value{
r (a list) containing details of the item and a plot
including details broken down by site_id (ICU)
}
\description{
Produces basic summary data for NHIC data items when passed NHIC data
in the form produced by the readItems() function. Calls seperate helper
functions for the different data types available.
}
\examples{
# Example call to inspect item function
# system.time(r <- inspect_item(rdt, ccdata.dict[["NIHR_HIC_ICU_0108"]]))
# Example plotting function
# ggplot(r$inspect$plot_data) +
#     geom_freqpoly(aes(x=val, y=..density.., colour=site_id),
#    binwidth=1)
}

