% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_ccDict.R
\name{is.ccDict}
\alias{is.ccDict}
\title{Confirms that the data dictionary is correctly formed}
\usage{
is.ccDict(d)
}
\arguments{
\item{d}{the object to be checked}
}
\value{
\code{TRUE}
}
\description{
Checks that the object is a list with items named by their HIC code
in the "NHIC_HIC_ICU_nnnn" format where n is [0,9]
}
\examples{
# is.ccDict(d)
}

