% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_HIC.R
\name{is.hic}
\alias{is.hic}
\title{Confirms that a data object matches the HIC specification}
\usage{
is.hic(d)
}
\arguments{
\item{d}{the object (data.table) to be checked}
}
\value{
\code{TRUE}
}
\description{
Checks that the object is a \code{data.table}, and then confirms that
it contains the 8 columns in the correct order
}
\examples{
# is.hic(d)
}
