% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_ICNARC.R
\name{is.valid.icnarc.code}
\alias{is.valid.icnarc.code}
\title{Checks ICNARC codes are valid}
\usage{
is.valid.icnarc.code(input)
}
\arguments{
\item{x}{Input string}
}
\value{
TRUE if valid, FALSE is not
}
\description{
Check if input is a valid ICNARC code
}
\examples{
# is.valid.icnarc.code('1.1')
# is.valid.icnarc.code('2.1.4.27.5')

}

