% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_ccDict.R
\name{load_ccDict}
\alias{load_ccDict}
\title{Loads the YAML field dictionary for ccdata}
\usage{
load_ccDict(path = NULL)
}
\arguments{
\item{path}{path to the fall back copy of the dictionary}
}
\value{
ccdata.dict
}
\description{
Tries to load the file from the ccdata package (assuming this is
installed) and if that fails then falls back to a local copy
Work with YAML file **NOT** with data.checklist
Extract file from ccdata package if possible
}
\examples{
# ccdata.dict <- load_ccDict() # loads using default values
# ccdata.dict <- load_ccDict("my/local/path") # alternative fall back path
}

