% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_ICNARC.R
\name{normalize.icnarc.code}
\alias{normalize.icnarc.code}
\title{Normalize an ICANRC code}
\usage{
normalize.icnarc.code(input)
}
\arguments{
\item{x}{Input string}
}
\value{
string with nornalized code, FALSE on error
}
\description{
Normalize an ICANRC code
}
\examples{
# normalize.icnarc.code ('1.01.02') returns 1.1.2

}

