% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_ICNARC.R
\name{parse.icnarc.code}
\alias{parse.icnarc.code}
\title{Split and label ICNARC diagnostic code}
\usage{
parse.icnarc.code(dt, dc.col, prefix = NULL)
}
\description{
Split and label ICNARC diagnostic code
}
