% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/self_test.R
\name{say_hello}
\alias{say_hello}
\title{Simple self-test}
\usage{
say_hello(x)
}
\arguments{
\item{x}{Your name}
}
\value{
Prints \code{x}
}
\description{
\code{say_hello} Prints a welcome message. This is the second '
paragraph and should briefly describe the function. All objects need a
title (the first line) and a description.
% <- see that percent sign, that's a latex comment that is
}
\details{
This is simply here for testing
}
\examples{
# say_hello("Steve")

}

