% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NHIC_utilities.R
\name{which_NHIC}
\alias{which_NHIC}
\title{Utilities that might be of use for working with NHIC data
Extract NHIC codes based on the value of a key:value pair in list}
\usage{
which_NHIC(k, v, f = NULL, dict_ = dict)
}
\arguments{
\item{k}{key}

\item{v}{value}

\item{f}{function returning logical}

\item{dict_}{NHIC dictionary from YAML file}
}
\description{
Utilities that might be of use for working with NHIC data
Extract NHIC codes based on the value of a key:value pair in list
}

