% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_apache_hr.R
\name{gen_apache_hr}
\alias{gen_apache_hr}
\title{Generates the APACHE Heart Rate score}
\usage{
gen_apache_hr(dt, input, output = NULL)
}
\arguments{
\item{dt}{data.table containing physiology data}

\item{input}{Heart Rate}

\item{output}{Column name for the result of computation}
}
\description{
Generates the APACHE Heart Rate score;
}
\examples{
# system.time(gen_apache_hr(ddata, hr = HR))
# table(ddata$apache_hr, useNA="always")
# ddata[hr > 40][sample(nrow(ddata[hr > 40]),20), .(hr)]
 @export
}

